services/system-service/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
cryptography==41.0.7

# JWT
pyjwt==2.8.0

# Redis (optional caching)
redis==5.0.1
hiredis==2.2.3

# Async Support
aiofiles==23.2.1
httpx==0.25.2
asyncio==3.4.3

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Environment & Config
python-dotenv==1.0.0

# Logging & Monitoring
structlog==23.2.0
colorama==0.4.6

# Testing (optional, for development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Data Validation & Serialization
email-validator==2.1.0
phonenumbers==8.13.26

# UUID Support
uuid==1.30

# CORS
fastapi-cors==0.0.6

# Health Check
fastapi-health==0.4.0

# Pagination
fastapi-pagination==0.12.13

# Rate Limiting
slowapi==0.1.9

# Background Tasks
celery==5.3.4
flower==2.0.1

# Utilities
python-json-logger==2.0.7
tenacity==8.2.3
